Pass the pid to the handlers for async processing
Instead of getting weird with registering global processes for the
running consumer we'll instead take a smarter approach of passing the
pid to the handler along with the message. The handler will do its async
processing and then call back to Consumer with the pid and message to
have the offset committed to Kafka.