Support connecting to Heroku Kafka (SSL)
This is a BIG change to Kaffe! We now support connecting to Heroku Kafka
(and any other Kafka setup that uses the same ENV variables for
endpoints and SSL).

Making this change in a reasonable way required moving all configuration
into the actual application configuration. Before this Kaffe was in a
weird place where some things were configured in the application
configuration and some things were expected to be supplied when starting
up the consumer/producer workers.

No more! Now all configuration is in the application configuration and
the consumer/producer workers pull everything they need from there.

Fixes #8