Clear subscribers on receving assignments
It appears we can get successive `assignments_received` without an
intervening `assignments_revoked`. I was not able to reproduce this
situation. It may be due to receiving messages out of order, or just
never receiving the latter message.

Brucke does clear existing subscribers when receiving either message.

Kaffe now follows this approach. In addition, an event is emitted when
this happens to provide some tracking.

Fixes #24