Add a function for producing a message list
The function uses Brod's support for producing lists like:

```
[{key, value}, {key, value} ...]
```

Also, the producer worker/process was used by all partition workers,
which made it a bottleneck in producing to Kafka. There is no longer a
separate process for the producer: delegation to Brod is used instead.

This does mean the round robin partitioning strategy is no longer
supported, as it was the only reason state was needed to be maintained
in the producer.

Finally, Kaffe is now a startable app. The producer client in Brod is
started if there is a producer configuration.

Updated the release to 1.0 due to the backward incompatibility.