Support different worker allocation strategies
The strategies are supported by variations in the worker name.

Currently, two strategies are supported:
- `:worker_per_partition`
- `:worker_per_topic_partition`

The `:worker_per_partition` strategy is the default and was the only
strategy previously supported by Kaffe. It allocates a worker per
partition across all topics.

The newly supported strategy is `:worker_per_topic_partition`. It
allocates a worker per topic/partition combination. This allows messages
to be read from the same partition in different topics concurrently.

The use of a single worker per partition across topics
(`:worker_per_partition`) is intended to prevent key collisions when
reading the same keys from a real time and batch topic simultaneously.

However, a collision in this configuration is not very likely.

Conversely, being able to read from the same partition across topics
simultaneously provides for better throughput and prevents
topic/partition starvation when a batch of messages take a long time.

Bumped the version to 1.2.0.

ENSI-224