Fix incorrect subscriber function declaration
Unfortunately the subscriber had a slightly incorrect function name
of commit_offset/5 instead of commit_offsets/5 (missing an "s" and
the end) and the worker was calling commit_offsets/5 which didn't exist
causing it to crash when a message handler intended for offsets to be
committed back to Kafka.

Moreover, the tests did not catch this since the modules injected during
the tests in place of the real subscriber had defined the function name
properly.