Run formatter, add some docs, and reorganize directory strucuture a bit
No functional changes (just yet!) in this commit. This is more to unify
some of the code formatting, reorganize the directory structure a bit,
and add some documentation on what each of the components are doing or
are used for since there are quite a few moving pieces and in some
places we rely on brod. A couple of structural changes are already planned
but this should break up the commits a bit.