Allow Kaffe Group Manager to recover when Kafka is unreachable
This commit addresses issue #93 which allows the Group Manager
to retry connecting group members to topics when Kafka is down.

The amount of retries and the retry interval are configurable with
Kaffe.Config.Consumer. Default is 3 retries every 10 seconds.

When retry options are pruned completely, Group Manager will no longer
be able to recover. This behaviour is expected, because it is most likely
that if reconnections are not possible in a set amount of time, intervention
is required. Retrying indefinitely does not make sense.